Updated opentrace-api to 0.31.0

+++ jdbc/src/main/java/io/opentracing/contrib/jdbc/JdbcPeer.java
@@ -5,6 +5,6 @@ public class JdbcPeer {
     public String name;
     public Integer ipv4;
     public String ipv6;
-    public Short port;
+    public Integer port;
 
 }
